Refactor db and health modules; reorganize svr structure
