+++ src/dbs/models.rs
@@ -1,3 +1,17 @@
+use std::sync::Arc;
+use surrealdb::Surreal;
+use surrealdb::engine::any::Any;
+
+pub type DbConnection = Arc<Surreal<Any>>;
 pub struct Database {
-    pub db: crate::dbs::connector::DbConnection,
+    pub db: DbConnection,
+}
+
+#[derive(Clone)]
+pub struct DbConfig {
+    pub endpoint: String,
+    pub namespace: String,
+    pub database: String,
+    pub username: String,
+    pub password: String,
 }
