+++ src/svr/health/mod.rs
@@ -1,0 +1,2 @@
+pub mod aggregator;
+pub mod models;
