+++ src/svr/mod.rs
@@ -1,3 +1,3 @@
+pub mod config;
 pub mod health;
 pub mod models;
-pub mod state;
