+++ src/svr/models.rs
@@ -1,24 +1,1 @@
-use serde::Serialize;
-#[derive(Serialize, Clone, Debug)]
-#[serde(rename_all = "lowercase")]
-pub enum HealthStatus {
-    Healthy,
-    Degraded,
-    Unhealthy,
-}
 
-// Single ComponentHealth struct (removing duplicates)
-#[derive(Serialize, Debug)]
-pub struct ComponentHealth {
-    pub name: String,
-    pub status: HealthStatus,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub message: Option<String>,
-}
-
-#[derive(Serialize)]
-pub struct SystemHealthResponse {
-    pub status: HealthStatus,
-    pub components: Vec<ComponentHealth>, // Changed from HashMap to Vec
-    pub timestamp: i64,
-}
